refactor: major structure change and refactor including: major tsconfig overhaul for a monorepo, breaking out server controllers, refactoring types, restructuring files and configs for monorepo
