--- models/users.ts
@@ -1,23 +1,0 @@
-import { Schema, model } from 'mongoose';
-
-const schema = new Schema({
-  email: { type: String, unique: true, required: true },
-  password: { type: String, required: true },
-  firstName: { type: String, required: true },
-  lastName: { type: String, required: true },
-  acceptTerms: Boolean,
-  role: { type: String, required: true },
-  verificationToken: String,
-  verified: Date,
-  resetToken: {
-    token: String,
-    expires: Date,
-  },
-  passwordReset: Date,
-  created: { type: Date, default: Date.now },
-  updated: Date,
-});
-
-const Users = model('Users', schema);
-
-export { Users };
