--- routes.ts
@@ -1,8 +1,0 @@
-import { Router } from 'express';
-import { userRoutes } from './controllers/Users';
-
-const routes = Router();
-
-routes.use('/users', userRoutes);
-
-export { routes };
