+++ server/_constants/roles.ts
@@ -1,6 +1,4 @@
-const roles = {
+export const roles = {
   ADMIN: 'admin',
   USER: 'user',
 };
-
-export { roles };
