+++ server/_helpers/hash.ts
@@ -1,0 +1,3 @@
+import { hashSync } from 'bcryptjs';
+
+export const hash = (password: string) => hashSync(password, 10);
