+++ server/_helpers/randomTokenString.ts
@@ -1,0 +1,3 @@
+import { randomBytes } from 'crypto';
+
+export const randomTokenString = () => randomBytes(40).toString('hex');
