+++ server/controllers/login.ts
@@ -1,0 +1,3 @@
+import type { Request, Response } from 'express';
+
+export const login = (_req: Request, res: Response) => res.send('login');
