+++ server/controllers/verifyEmail.ts
@@ -1,0 +1,3 @@
+import type { Request, Response } from 'express';
+
+export const verifyEmail = (_req: Request, res: Response) => res.send('verify-email');
