+++ server/database/connection.ts
@@ -1,12 +1,12 @@
 require('dotenv').config();
-import mongoose from 'mongoose';
+import {connect, connection} from 'mongoose';
 
-export const connect = () => {
-  mongoose.connect(
+export const dbConnect = () => {
+  connect(
     `mongodb+srv://${process.env.MONGO_USER}:${process.env.MONGO_PASSWORD}@${process.env.MONGO_CLUSTER_NAME}/${process.env.MONGO_TABLE_NAME}?retryWrites=true&w=majority`
   );
 
-  const db = mongoose.connection;
+  const db = connection;
   db.on('error', console.error.bind(console, 'connection error:'));
   db.once('open', () => console.log('we are connected to mongo db'));
 };
