+++ server/index.ts
@@ -4,8 +4,8 @@ import bodyParser from 'body-parser';
 import cookieParser from 'cookie-parser';
 import cors from 'cors';
 
-import { connect } from './database/connection';
-import { routes } from './routes';
+import { dbConnect } from './database/connection';
+import { router } from './controllers/router';
 
 const app = express();
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -13,20 +13,19 @@ app.use(bodyParser.json());
 app.use(cookieParser());
 
 // allow cors requests from any origin and with credentials
-app.use(cors({ origin: (origin, callback) => callback(null, true), credentials: true }));
+app.use(cors({ origin: (_origin, callback) => callback(null, true), credentials: true }));
 
-app.use(function (req, res, next) {
+app.use(function (req, _res, next) {
   console.log(`${req.method} request for ${req.url}`);
   next();
 });
 
 if (process.env.DB_CONNECTION === 'true') {
-  connect();
+  dbConnect();
 }
 
 // api routes
-app.get('/', (req, res) => res.send('Welcome to the api server'));
-app.use('/api', routes);
+app.use('/api', router);
 
 const port = process.env.PORT || 5000;
